Clean up bit. Break out startApp into separtate module to resuse in ui test file.
