+++ script/start-app.ts
@@ -1,0 +1,18 @@
+import { run } from './run'
+
+export function startApp() {
+  const runningApp = run({ stdio: 'inherit' })
+  if (runningApp == null) {
+    console.error(
+      "Couldn't launch the app. You probably need to build it first. Run `yarn build:dev`."
+    )
+    process.exit(1)
+    return
+  }
+
+  runningApp.on('close', () => {
+    process.exit(0)
+  })
+
+  return runningApp
+}
